ausk/backend/requirements.txt
# FastAPI and async dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Redis
redis[hiredis]==5.0.1

# Validation & Config
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
python-dotenv==1.0.0

# Email & Templates
Jinja2==3.1.2
aiosmtplib==3.0.1

# Task Queue
celery==5.3.4

# MFA & QR Codes
pyotp==2.9.0
qrcode[pil]==7.4.2

# HTTP Client
aiohttp==3.9.1

# Stripe Billing
stripe==7.7.0

# YAML Processing
PyYAML==6.0.1

# Data Room Dependencies
cryptography==41.0.7
tika==2.6.0
python-docx==0.8.11
python-pptx==0.6.21

# ML Dependencies (for embedder service)
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
grpcio==1.59.3
grpcio-tools==1.59.3

# Vector database
pgvector==0.2.4